Move collect_source_model_paths in logictree [skip hazardlib]
